requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
pymongo==4.6.1
faker==22.0.0

# Connection Pooling
sqlalchemy==2.0.25

# Environment Variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# For API Layer (optional - for future integration)
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==42.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
